pyproject.toml
[build-system]
requires = ["setuptools>=67", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mlops-agent"
version = "0.1.0"
description = "LangGraph-powered MLOps agent for Azure ML workspace management"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    # LangChain / LangGraph
    "langchain>=0.3",
    "langchain-anthropic>=0.3",
    "langchain-core>=0.3",
    "langgraph>=0.3",
    "langgraph-checkpoint-sqlite>=1.0",

    # FastAPI backend
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    "python-dotenv>=1.0",
    "pydantic>=2.0",

    # Chainlit frontend
    "chainlit>=2.0",
    "httpx>=0.27",

    # Azure SDKs
    "azure-ai-ml>=1.13",
    "azure-identity>=1.15",
    "azure-core>=1.30",
    "msal>=1.27",
    "requests>=2.31",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-mock>=3.12",
    "ruff>=0.4",
    "mypy>=1.10",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]

[tool.ruff]
line-length = 100
target-version = "py310"
select = ["E", "F", "I", "UP"]
